requirements.txt
transformers==4.28.1
sentencepiece==0.1.98
underthesea==6.2.0
gensim==4.3.1
numpy==1.24.2
pandas==2.0.0
matplotlib==3.7.1
torch==2.0.0
bcrypt==4.0.1
beanie==1.11.0
dynaconf==3.1.8
fastapi==0.88.0
fastapi-jwt-auth==0.5.0
fastapi-mail==1.1.5
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
Jinja2==3.1.2
motor==2.5.1
orjson==3.8.3
passlib==1.7.4
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
pymongo[srv]==3.12.3
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
starlette==0.22.0
typing_extensions==4.4.0
ujson==5.6.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4